LPS-85088 Expand LanguageImplTest to multiple classes
